requirements.txt
Django==1.11
mysqlclient
WeasyPrint==0.42
django-simple-history==1.9.0


